More invoice debug. (in accounting added test invoice). Legacy - fix (non static) + shipping is float.

+++ wp-content/plugins/fresh/includes/class-fresh-order.php
@@ -80,7 +80,7 @@ class Fresh_Order {
 		return null;
 	}
 
-	public function getShippingFee() : int
+	public function getShippingFee() : float
 	{
 		$sql2 = 'SELECT meta_value FROM `wp_woocommerce_order_itemmeta` WHERE order_item_id IN ( '
 		        . 'SELECT order_item_id FROM wp_woocommerce_order_items WHERE order_id = ' . $this->order_id
